write lots of exams; fixes to Canvas
When there are several quizzes in a Canvas export, one .exam file is
written for each of them.

Images are copied out to a directory with the same name as the .exam
file.

The multiple-choice question types have "shuffle choices" turned on.

Canvas equation images are replaced with delimited LaTeX.